requirements.txt
streamlit>=1.25.0  # Ensure the latest compatible Streamlit version
Pillow>=9.5.0
transformers==4.34.0  # Ensure compatibility with Python 3.12.1
faiss-cpu
numpy>=1.25.0
python-dotenv>=1.0.0
google-generativeai>=0.1.1  # Check compatibility with gemini API
torch>=2.0.0  # Ensure compatibility with the version of `transformers`
